refactor: extract magic numbers as named constants; add trailing-whitespace lint check
Replace scattered hardcoded numeric literals with named constants across several
modules, improving readability and ensuring a single definition for each tuneable
value. Also add a new lint check (Check 18) that enforces no trailing whitespace
in source and documentation files, and clean up the existing violations it found.

Source changes:

- at-ps2: extract ATPS2_INIT_CHECK_MS (200), ATPS2_ID_RETRY_LIMIT (2), and
  ATPS2_DETECT_MAX_ATTEMPTS (5) from keyboard_interface_task(); removes the
  last hardcoded timeouts and retry counts from the AT/PS2 initialisation path

- command_mode/ws2812: extract CMD_MODE_BRIGHTNESS_MAX and use WS2812_BRIGHTNESS_MAX
  in place of the literal 10 used for LED brightness range clamping; both modules
  now reference the same conceptual ceiling through their respective headers

- uart: extract UART_DMA_BACKOFF_MIN_US (1) and UART_DMA_BACKOFF_MAX_US (1024) from
  the exponential-backoff loop in wait_for_queue_space(); header comment updated to
  clarify printf/puts are not safe in IRQ context directly (use LOG_* macros)

- keymaps: add early return for KC_NO before shift-override lookup, preventing
  redundant processing on unmapped key positions

- hid_interface: restructure hid_send_report() to capture tud_hid_n_ready() result
  into a local bool rather than early-returning, allowing the log path to remain
  reachable regardless; send is gated with ready && tud_hid_n_report(...)

- keyboards (all): cast keymap_layer_count initialiser to (uint8_t) explicitly,
  suppressing implicit narrowing from size_t; also remove trailing spaces from
  ASCII art layout comments in keyboard.c/.h files

- amiga: correct the handshake timing comment to reflect the actual ~15.5ms
  measured duration rather than the target 85µs

Tools:

- lint.sh: add Check 18/18 — trailing whitespace detection; source files (.c/.h)
  fail on any trailing whitespace; markdown files fail on single trailing space
  (double trailing spaces are preserved as intentional Markdown hard line breaks);
  all existing check counters updated from [N/17] to [N/18]

- test-enforcement.sh: add test_lint_trailing_whitespace() covering four scenarios
  (source violation, markdown single-space violation, markdown hard-break pass,
  clean file pass); fix test_lint_pio_irq_dispatcher_correct() scaffold to include
  license header and ringbuf_reset() so it passes all 18 checks as intended

- tools/README.md: document Check 18, update check count to eighteen, update
  example output to show [1/18]/[2/18]

Docs:

- Strip single trailing spaces from ~40 markdown files in docs/ identified by the
  new lint check; double trailing spaces (Markdown hard line breaks) are preserved

+++ src/keyboards/amiga/a500/keyboard.c
@@ -25,7 +25,7 @@
 // clang-format off
 
 /* Commodore Amiga 500/2000 Keyboard Layout
- * 
+ *
  * ,---. ,------------------------. ,------------------------.
  * |Esc| |F1  |F2  |F3  |F4  |F5  | |F6  |F7  |F8  |F9  |F10 |
  * `---' `------------------------' `------------------------'
@@ -58,7 +58,7 @@ const uint8_t keymap_map[][KEYMAP_ROWS][KEYMAP_COLS] = {
 };
 
 /* Layer count - automatically calculated from keymap_map array size */
-const uint8_t keymap_layer_count = sizeof(keymap_map) / sizeof(keymap_map[0]);
+const uint8_t keymap_layer_count = (uint8_t)(sizeof(keymap_map) / sizeof(keymap_map[0]));
 
 /*
  * Key mapping notes for Commodore Amiga 500/2000:
